requirements.txt
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0
requests==2.31.0
